scraper-google-email/requirements.txt
selenium>=4.24.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dotenv>=1.0.1
webdriver-manager>=4.0.2
